pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lineindex"
description = "Fast line-based random access to large text files with optional compression"
authors = [
    {name = "Miroslav Lžičař", email = "mireklzicar@seznam.com"},
]
readme = "README.md"
requires-python = ">=3.6"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Topic :: Text Processing :: Indexing",
    "Topic :: System :: Archiving :: Compression",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
]
dynamic = ["version"]
dependencies = [
    "numpy>=1.20.0",
    "tqdm>=4.50.0",
]

[project.optional-dependencies]
compression = ["python-idzip>=0.3.3"]
lint = ["flake8", "flake8-pyproject", "black", "isort"]
test = ["pytest", "mypy", "pytest-cov"]
docs = ["sphinx", "sphinx-rtd-theme", "myst-parser"]
dev = ["lineindex[compression,lint,test,docs]", "bump2version"]

[project.scripts]
lineindex = "lineindex.cli:main"

[tool.setuptools]
packages = ["lineindex"]
package-data = {"lineindex" = ["data/*.txt"]}

[tool.flake8]
max-line-length = 99
extend-ignore = "E203,D401,D400,D107,D105,D104,E302,W293"
exclude = ["build", "dist", "*.egg-info", ".git", ".tox", "__pycache__", "docs"]

[tool.black]
line-length = 99
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 99

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"